Disable Subversion support
I don't use svn myself and never heard of anyone using it with Integrity,
so I am disabling it for now and making Integrity a lit bit more git-user
friendly by providing "master" as the default branch and not displaying the SCM
dropdown on the creation form.

If you do care about svn, please feel free to revert this commit and report us
how it go. I'd be happy to re-enable it as soon as I known it is actually
working and useful for someone.