first pass on auto-branch feature
The idea is to create a project automatically if we get a post-receive hook
for a branch that is not being watched.

Say we have a project named "Integrity" tracking the master branch. I am
currently working on a new feature in a topic branch named "autobranch".

  git checkout -b autobranch
  # hack hack hack
  git commit
  git push origin autobranch

Integrity will create a project named "Integrity (autobranch)" and build
that branch for me.

It is not very useful for Integrity itself which has a relatively fast and
easy to run test suite. It is, however, very convenient when dealing with a
slow-ass and setup-heavy test-suite to be able to literally throw the code
at Integrity and let it take it from there.

As a plus, it makes topic branches test status easily dereferenceable.

"Please review and merge this branch; here is the GitHub compare view and the
build URL"