Deal with incompatible psych YAML parser
The new psych YAML parser is not 100% compatible with the older
parser.  If any gem requests psych, it alters all future YAML
parsing. In my environment bundler-1.0.18 was the culprit.

checkout.rb expects `committed_at` to be a string, but psych
automatically parses it into a `Time`.  This patch should make it work
with either parser.