Moved checking out process into "build".
Build will now be considered started, and notifications will be sent,
prior to any commands being executed on the build repository.

This will have the effect of properly marking builds that fail due to
exceptions raised during the build repository preparation process
as failed; previously, such builds were considered not started.

This change has the potentially undesired effect of crediting the
time it takes to prepare the build repository to the build process.
However, some may want this behavior, and build process itself may
perform setup tasks that may not be desired as part of build time.
It is still possible to time any interesting part of build manually
by using the time utility or any other means.