Fix do_sqlite3 dependency for ruby 1.9.3-preview1.
do_sqlite3 < 0.10.7 uses DateTime.new! which does not exist in
ruby 1.9.3-preview1 (it was removed somewhere in 1.9 line).
When running on 1.9, require 0.10.7.

For 1.8, 0.10.7 seems to work fine, at least in tests.
However due to massive amounts of spam generated by 0.10.5+
with datamapper 1.0.x, keep the version at 0.10.2 as it was
previously.