Add missing queue attribute/column to delayed_jobs model/table, respectively.
delayed_job 3.0.2 has provision for this attribute. delayed_job 3.0.5
specifies a default, and therefore fails without this attribute being
declared in the model.

Missing attribute and column was a bug.

I cannot tell if delayed_job 3.0.2 actually functioned without this
column in the database, especially if no queues were declared.

This commit does not add the column to existing databases, on assumption
that the installed pool using this configuration is small, the change
is easy to make on any given dbms and difficult to implement in a
generic fashion.

Further discussion can be found in #221.